# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface YoutubeVideo {
  videoId: string;
  title: string;
  description: string;
  thumbnailUrl: string;
  publishedAt: string;
  channelTitle: string;
}

export interface HealthStatus {
  status: string;
}

export type CreateStorySubmissionInputPreferredContactMethod = typeof CreateStorySubmissionInputPreferredContactMethod[keyof typeof CreateStorySubmissionInputPreferredContactMethod];


export const CreateStorySubmissionInputPreferredContactMethod = {
  email: 'email',
  phone: 'phone',
  either: 'either',
} as const;

export interface CreateStorySubmissionInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  phone?: string;
  /** @minLength 1 */
  story: string;
  preferredContactMethod: CreateStorySubmissionInputPreferredContactMethod;
}

export type StorySubmissionPreferredContactMethod = typeof StorySubmissionPreferredContactMethod[keyof typeof StorySubmissionPreferredContactMethod];


export const StorySubmissionPreferredContactMethod = {
  email: 'email',
  phone: 'phone',
  either: 'either',
} as const;

export interface StorySubmission {
  id: string;
  fullName: string;
  email: string;
  phone?: string | null;
  story: string;
  preferredContactMethod: StorySubmissionPreferredContactMethod;
  createdAt: string;
}

export interface CreateContactMessageInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactMessage {
  id: string;
  name: string;
  email: string;
  message: string;
  createdAt: string;
}