# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /youtube/latest:
    get:
      operationId: getYoutubeLatest
      tags: [youtube]
      summary: Get latest YouTube videos
      description: Returns the latest videos from The Mouthpiece YouTube channel
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YoutubeVideo"
        "502":
          description: YouTube API error
...
components:
  schemas:
    YoutubeVideo:
      type: object
      properties:
        videoId:
          type: string
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        publishedAt:
          type: string
        channelTitle:
          type: string
      required:
        - videoId
        - title
        - description
        - thumbnailUrl
        - publishedAt
        - channelTitle
...